separate sheets operations
