+++ quadratic-core/src/controller/operations/mod.rs
@@ -9,3 +9,4 @@ pub mod clipboard;
 pub mod formatting;
 pub mod import;
 pub mod operation;
+pub mod sheets;
