+++ quadratic-core/src/controller/user_actions/mod.rs
@@ -6,4 +6,5 @@ pub mod cells;
 pub mod clipboard;
 pub mod formatting;
 pub mod import;
+pub mod sheets;
 pub mod undo;
